Use smart constructors in Haskell GDPR example.
This means we lose the record syntax when constructing the Contact. That
makes us depend on argument order and thereby opens up, in practice, the
possibility of mixing up name and email. I was not able to find a way of
writing smart constructors like dumb record constructors, such that it's
possible to use record syntax, while still being usable as a 'function'.

The upside of this approach is that it forces the API consumer to choose
between limited options, and disables the ability to construct a Contact
where their status as EU residents hasn't been at least considered.